requirement.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens
attrs==24.2.0
audioread==3.0.1
av==12.3.0
azure-cognitiveservices-speech==1.41.1
bitsandbytes==0.44.1
blobfile==3.0.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chattts==0.2.1
click==8.1.7
cn2an==0.5.22
coloredlogs==15.0.1
contourpy==1.3.0
ctranslate2==4.5.0
cycler==0.12.1
Cython==3.0.11
dataclasses==0.6
datasets==3.1.0
decorator
dill==0.3.8
einops==0.8.0
einx==0.3.0
encodec==0.1.1
evaluate==0.4.3
evdev==1.7.1
executing
fairscale==0.4.13
fastapi==0.115.4
faster-whisper==1.0.3
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.54.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-texttospeech==2.21.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.10
iniconfig==2.0.0
ipython
jedi
Jinja2==3.1.4
jiwer==3.0.5
joblib==1.4.2
keyboard==0.13.5
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
-e git+https://github.com/meta-llama/llama3.git@11817d47e1ba7a4959b025eb1ca308572e0e3963#egg=llama3
llama_models==0.0.47
llama_stack==0.0.47
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
noisereduce==3.0.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.3.3
onnxruntime==1.20.0
openai-whisper==20240930
opencc-python-reimplemented==0.1.7
packaging==24.1
pandas==2.2.3
parso
peft==0.13.2
pexpect
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
proces==0.1.7
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess
pure-eval
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pybase16384==0.3.7
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments
pynput==1.7.7
pyparsing==3.2.0
pytest==8.3.3
pytest-runner==6.0.1
pytest_httpserver==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-xlib==0.33
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.1.0
six
sniffio==1.3.1
SoundCard==0.4.3
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
stack-data
starlette==0.41.2
sympy==1.13.1
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.4.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
tqdm==4.66.5
traitlets
transformers==4.46.2
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
vector-quantize-pytorch==1.19.2
vocos==0.1.0
wcwidth
Werkzeug==3.1.3
wheel==0.44.0
xxhash==3.5.0
yarl==1.17.1
zhconv==1.4.3
